discord-webhook-service/requirements.txt
pydantic
fastapi
requests
python-dotenv